# interface/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API containing the two challenges pertaining to the SA programming challenge",
        "title": "Solutions Architect Programming Challenge",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/algorithm/fizzbuzz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Run classic FizzBuzz algorithm on the interval 1-100",
                "responses": {}
            }
        },
        "/algorithm/intervals": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Run the algorithm to get intervals, receiving includes and excludes arrays",
                "parameters": [
                    {
                        "description": "{ includes: [[10, 100],[200, 300],[400, 500]], excludes: [[95, 205],[410, 420]]}",
                        "name": "intervals",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.intervalsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.intervalsRequest": {
            "type": "object",
            "properties": {
                "excludes": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "includes": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}